# Remove `RecursiveArrayTools.ArrayPartition` -- do partitioning manually